Notifications: use environment variable for state
We can use the environment variable to keep the state of the toggles
when we leave the view. This is really messy right now, but I hope
to make the notifications object an array so we can iterate through
and display them dynamically.

Note: I needed to add the  to make it so that,
when you flicked a toggle, you *weren't* brought back a view. Super
weird bug

Also: I had to change the PlantObject id to be `=` to the UUID() instead
of a function call. This is because the function call was being called
every time the ID was accessed, ruining the whole point of the ID.